binding

+++ OstManSysMVVM/Model/Resident.cs
@@ -11,7 +11,7 @@ namespace OstManSysMVVM.Model
         public int ResidentID { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public long PhoneNumber { get; set; }
+        public int PhoneNumber { get; set; }
         public string EmailAddress { get; set; }
         public DateTime DateOfBirth { get; set; }
         public string Type { get; set; }
